code golf

+++ src/main/java/java8/util/ArrayDequeSpliterator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2013, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2016, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -110,7 +110,7 @@ final class ArrayDequeSpliterator<E> implements Spliterator<E> {
         if (n < 0) {
             n += getData(deq).length;
         }
-        return (long) n;
+        return n;
     }
 
     @Override
